Polish sql parser package
