+++ shardingsphere-sql-parser/shardingsphere-sql-parser-engine/src/main/java/org/apache/shardingsphere/sql/parser/api/SQLParserEngine.java
@@ -19,8 +19,8 @@ package org.apache.shardingsphere.sql.parser.api;
 
 import com.google.common.cache.LoadingCache;
 import org.antlr.v4.runtime.tree.ParseTree;
-import org.apache.shardingsphere.sql.parser.core.cache.ParseTreeCacheBuilder;
-import org.apache.shardingsphere.sql.parser.core.parser.SQLParserExecutor;
+import org.apache.shardingsphere.sql.parser.core.database.cache.ParseTreeCacheBuilder;
+import org.apache.shardingsphere.sql.parser.core.database.parser.SQLParserExecutor;
 
 /**
  * SQL parser engine.
